+++ CHANGES.rst
@@ -11,11 +11,21 @@ bugs:
 
 - Fixed various bugs with the ``@arg_botcmd`` decorator (`#516 <https://github.com/gbin/err/pull/516>`_)
 - Fixed warn_admins() on Telegram
+- Slack ACLs now properly check against usernames starting with `@`
+- Slack identifiers can now be built from a bare `#channel` string (without a username part)
+- Slack identifiers can now be built from a `<#C12345>` or `<@username>` string (the webclient formats them like this automatically when chatting with the bot)
 
 other:
 
 - Removed some dead code
 - Removed deprecated bare_send and invite_to_room bot methods
+
+v3.1.1 (2015-10-26)
+-------------------
+
+bugs:
+
+- fix regression on !help (thx kromey)
 
 
 v3.1.0 (2015-10-22)